mercurial: tidy up eden import helper mononoke detection
Summary:
We've been using a hard coded list to determine whether
mononoke is available, and that list is falling behind the current
state of our backend migration.

This commit removes the hard coded list in favor of testing
`remotefile.reponame`.  If that is configured then it holds
the mononoke reponame.

I'm making the assumption that it being set implies that mononoke
is available for that repo.  That may not be 100% true, but
it appears to be effectively true for the intersection of repos
known to `fbclone` and the migrated set of repos.

If this code decides that mononoke is supported, the behavior
in EdenFS is to then attempt to use the SCS to fetch tree
data using the returned `repo.name`.  That appears to be the
only way that this information is used today.

Reviewed By: quark-zju

Differential Revision: D23214471

fbshipit-source-id: 17b6475b891df5423dca0c18ddae0838795f713a