eden: fixup debug level for SpawnedProcess
Summary:
I switched these from ERR to DBG2 thinking that would
take them out of the log by default, but we log DBG2 by default
so that didn't have the desired result.

This changes the log level to DBG6 which is an arbitrary log
level that isn't included by default.

Reviewed By: genevievehelsel

Differential Revision: D23467738

fbshipit-source-id: dd0c75f86318ece27313c237938f24f55758eec1